@@ -12,24 +12,26 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/common": "4.0.0",
-        "@angular/compiler": "4.0.0",
-        "@angular/core": "4.0.0",
-        "@angular/forms": "4.0.0",
-        "@angular/http": "4.0.0",
-        "@angular/platform-browser": "4.0.0",
-        "@angular/platform-browser-dynamic": "4.0.0",
-        "@angular/router": "4.0.0",
+        "@angular/animations": "5.2.0",
+        "@angular/common": "5.2.0",
+        "@angular/compiler": "5.2.0",
+        "@angular/core": "5.2.0",
+        "@angular/forms": "5.2.0",
+        "@angular/http": "5.2.0",
+        "@angular/platform-browser": "5.2.0",
+        "@angular/platform-browser-dynamic": "5.2.0",
+        "@angular/platform-server": "5.2.0",
+        "@angular/router": "5.2.0",
         "core-js": "2.4.1",
         "foundation-sites": "6.3.1",
         "jquery": "3.2.1",
         "ng2-translate": "5.0.0",
-        "rxjs": "5.1.0",
+        "rxjs": "5.5.6",
         "zone.js": "0.8.4"
     },
     "devDependencies": {
         "@angular/cli": "1.0.0",
-        "@angular/compiler-cli": "4.0.0",
+        "@angular/compiler-cli": "5.2.0",
         "@types/jasmine": "2.5.38",
         "@types/node": "6.0.60",
         "codelyzer": "2.0.0",
@@ -44,6 +46,6 @@
         "protractor": "5.1.0",
         "ts-node": "2.0.0",
         "tslint": "4.5.0",
-        "typescript": "2.2.0"
+        "typescript": "2.4.2"
     }
 }
